refactor(structure): move article builder to a more logical spot
